+++ src/main/java/onlyoffice/OnlyOfficeConvertServlet.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -250,6 +250,7 @@ public class OnlyOfficeConvertServlet extends HttpServlet {
                 copy.setMediaType(documentManager.getMimeType(newName));
 
                 attachmentManager.saveAttachment(copy, null, inputStream);
+                attachmentUtil.setCollaborativeEditingKey(copy.getLatestVersionId(), null);
 
                 return copy.getLatestVersionId();
             }