+++ src/main/java/onlyoffice/conditions/IsOfficeFileAttachment.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -77,7 +77,7 @@ public class IsOfficeFileAttachment implements Condition {
         } else {
             if (accessView
                     && documentManager.isViewable(fileName)
-                    && !(accessEdit && (documentManager.isEditable(fileName) || documentManager.isFillable(fileName)))
+                    && !(accessEdit && documentManager.isEditable(fileName))
             ) {
                 return true;
             }