+++ src/main/java/onlyoffice/sdk/manager/url/UrlManagerImpl.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -188,6 +188,12 @@ public class UrlManagerImpl extends DefaultUrlManager implements UrlManager {
                 nameIcon + ".ico",
                 UrlMode.ABSOLUTE
         );
+    }
+
+    public String getUsersInfoUrl() {
+        String usersInfoUrl = getConfluenceBaseUrl(false) + API_SERVLET + "?type=users-info";
+
+        return usersInfoUrl;
     }
 
     private String getConfluenceBaseUrl(final Boolean inner) {