+++ CHANGELOG.md
@@ -5,7 +5,15 @@
 - core of the plugin has been moved to com.onlyoffice.docs-integration-sdk (https://github.com/ONLYOFFICE/docs-integration-sdk-java)
 - improved connection settings validation
 - improved history servlet security
+- setting authorization header on settings page
+- user image in editor
+- filling pdf
 - translations (zh_CN)
+
+## Changed
+- improved date formatting in history changes
+- default conversion format (from docxf to pdf instead oform)
+- remove filling for oform
 
 ## 4.4.0
 ## Added